// sharing-frontend-RWN/packages/shared/src/api/generated/schemas/activityInput.ts
/**
 * Generated by orval v7.8.0 🍺
 * Do not edit manually.
 * Travel Planner API
 * API documentation for the Travel Planner application
 * OpenAPI spec version: 1.0.0
 */
import type { ActivityInputType } from './activityInputType';

export interface ActivityInput {
  title: string;
  description?: string;
  type: ActivityInputType;
  startTime: string;
  endTime?: string;
}